Wed, Sep 9, 2026, 10:53 PM +03:00

+++ avg.c
@@ -24,6 +24,6 @@ int main (void)
 
     }
 
-    printf("highest is %i\n", highst);
+    printf("highest is %f\n", (float) highst);
 
 }
